rok_bot/requirements.txt
opencv-python
numpy
Pillow
pure-python-adb
pytesseract
pyyaml
